Move types into dedicated implementation files
