+++ src/index.ts
@@ -1,1 +1,3 @@
-export { type AsyncModalProps, type AsyncModalState, useAsyncModalState } from './useAsyncModalState'
+export { type AsyncModalProps } from './AsyncModalProps'
+export { type AsyncModalState } from './AsyncModalState'
+export { useAsyncModalState } from './useAsyncModalState'
